+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Product Backlog
+* [High] Trim polyfill to only needed features (exclude not working)
 * [High] Build SHYnet page
 * [Low] Add method to push newsletters from admin
 * [Low] Embed images in order confirmations and newsletters using nodemailer